unify: require sudo earlier
I keep forgetting to type sudo and then it times out. Now, it'll request
it way earlier - right after it swaps branches and stuff, but before it
runs `nix flake update`. This will mean you have to type sudo more often
if `unify` often fails for you, but I think it's worth it.