flake: remove llakaLib, add non-flake entrypoint
With this, we finally no longer rely on llakaLib for anything - all our
functions are defined locally.

We also simplify the package definition process as much as possible.
While `lib.filesystem.packagesFromDirectoryRecursive` is very nice for
UX, it's a lot more of a black box. If we're okay with manually listing
the files, we can achieve similar logic, with only `callPackageWith` and
a little `let` recursion.

Finally, non-flake users are now first-class citizens! Just pass an
instantiated `pkgs` set into `./packages/default.nix`, and you'll get
all the features flake users would get. This should really be documented
in the README - but I'll leave that to future me.