fuiska: read from duplicate inputs correctly
Flakes are stupid. If you have duplicate inputs, they suffix the
duplicates with `_`. So if you have two home-managers in your
flake.lock, they become `home-manager` and `home-manager_2`. The stupid
part? Your inputs don't take precedence. The non-underscore one might be
a random transitive dependency.

I thought I'd solved this, but it turns out I'd actually been reading
from the keys, not the values. But if I passed the values, fuiska would
print `home-manager_2` when it had an update, instead of just
`home-manager`. To fix this, we pass the real name and the location in
the file separately.