fight: prevent false positives
Home-manager has a branch now called `update/master`. This is causing
issues with my `git ls-remote` usage, since it matches on anything after
a slash. To avoid this, we do extra logic if we got more than one ref
from `git ls-remote`. If that last ref's doesn't end with `^{}`, we can
assume we're not dealing with annotated tags, and instead just an
ignorable duplicate. I can't guarantee this will work for all inputs,
but it fixes the home-manager case, so I'll take what I can get.