imanpu/qwesu: add Channel support
As always, here's your government mandated regex explainer. We'll use
this example URL:
https://releases.nixos.org/nixos/unstable/nixos-25.11pre882227.01f116e4df6a/nixexprs.tar.xz

And, here's the full regex if you don't want to keep checking back:
https:\/\/releases\.nixos\.org\/(((?:nixos\/)?[^\/]*\/[^-]*-\d+\.\d+)[^\/]*)

First, we grab the beginning of the url with
`https:\/\/releases\.nixos\.org\/`. If you can't handle dot escaping,
you won't like what's coming.

Next, we start a big capturing group. This will contain the "key" of the
current revision, so we can compare it to the key we get from querying
the s3 db later. It'll look something like:
`nixos/unstable/nixos-25.11pre882227.01f116e4df6a`

We then start a second capturing group. This will be within the key one,
and will contain the prefix that we'll be using to search the S3
database, not all the revision stuff at the end. It'll look something
like: `nixos/unstable/nixos-25.11`.

Now, need to parse the next few slash-separated strings. In a perfect
world, we could just match for slash-suffixed strings - but
`nixpkgs-unstable` channels will look like `nixpkgs/nixpkgs-25.11`. As a
dirty workaround, we match for an optional `nixos/` with `(?:nixos\/)?`.
Finally, we match for one slash-suffixed string with `[^\/]*\/`.

Next, we need to include the actual channel name. On unstable branches,
the channel name is followed by `pre` - on stable branches, it's
followed by a period. Rather than doing conditional logic here, we just
assume the channel will follow basic nixos versioning rules. Start with
everything until a dash with `[^-]*`, then a literal `-`, and finally
two numbers separated by a period with `\d+\.\d+`.

At this point, we now end the prefix capturing group. We're still in the
key capturing group, though - and we need to match all the hash
nonsense. This is actually really easy - just match until the slash with
`[^\/]*`, and end the capturing group. We're done!