fuiska, hue: improve directory logic
New behavior here. If an arg for `-d` is passed, we use it, and error if
it's invalid. However, if you don't pass a directory, we now use the
current dir, _if_ it passes hue tests. We only fall back to the default
dir if the current dir is invalid.

We also actually print the dir that's being used on execution, with some
pretty logic, so `../` gets expanded, and `/home/user/foo/bar` gets
turned into `~/foo/bar`.